GL: drm: Introduce OpenGL/ES OSD
Use GLES to compose the OSD.
This is based on the work of vdr# user louisbraun https://github.com/louisbraun/softhddevice-openglosd
which is based on OpenGL. In order to use that piece of code on ARM platform, which don't support OpenGL
i adapted the code to use OpenGL/ES 2.0.

I also committed some fixes from https://github.com/jojo61/vdr-plugin-softhdcuvid/commits/master/openglosd.cpp,
mainly the support for DrawPixel and the freetype fixes.

One difference is, that we don't use the VDPAU interop functionality anymore. We bind a texture to the
output framebuffer and simply draw the (ready) composed OSD onto that texture. eglSwapBuffers() swaps
the GL's front- and backbuffer. Drm gets the buffer from our gbm surface and creates a drm framebuffer
for it, which then can be used for the plane setting.

The main difference to both is the handling of DrawText. This is now done with a texture atlas.
As soon as a new font is created, the openglosd creates a texture for each fontheight, wich holds all
characters of the the Basic Latin (ASCII) and Latin-1 Supplement character sets.
This should be an improvement, because we do not need to create the texture and do glDrawArrays for
every single character anymore. We just need to bind the texture (because we created it already before)
and do a glDrawArrays for each DrawText call, which is one call per string.

Usage:
If the following requirements are installed, they are autodetected durign build process and the plugin
will automatically get GLES support. Setting the environment variable GLES to 1 or 0 during compilation
forces or disables building the plugin with GLES.

Requirements:
- GLES userspace lib and headers (mesa)
- freetype2
- glm

Signed-off-by: Andreas Baierl <ichgeh@imkreisrum.de>