drm: Rework SetPropertyRequest()
It's probably expensive to iterate to the properties just before the atomic commit.
So once we have the plane_id, we fill an array with its properties an can just read
this array later.

We introduce a struct with the planes' preoperties and move the plane_id into this
struct.

Signed-off-by: Andreas Baierl <ichgeh@imkreisrum.de>